requirements.txt
flask
mysql-connector-python
pandas
requests
scipy
prophet
numpy